TCPHandler: Move constructor into the module and add missing headers
